Adding requires_arc = 'true' to pass trunk spec validation.
Motivation:
When using 0.36.0.beta.1 the podspec validates without any problems. But
it seems like trunk is not using the same version and the following
error is given when trying to push to trunk:
A value for `requires_arc` should be specified until the migration to a
`true` default

Modifications:
Added requires_arc = 'true' to the podspec to see if this will allow it
to be pushed.