Checking for already received patches before checking backup.
Motivation:
Currently we are checking if we need to restore from a backup copy would
be required if a patch message was never recieved by the server. This is
done before we check if we already have the patch in question. I believe
that this is wrong and that we should be checking if we have the patch
already first, giving us the chance to discard it, before going down the
backup route.

I only noticed this when using the server-wildfly module and never with
the server-netty module. This might just be bad luck that it never
showed up.